deleted unnecessary interface for JwtService
